mv shared types to `types` package
