+++ packages/app-staking/src/library/QRSignPrompt/index.tsx
@@ -14,9 +14,9 @@ import { getStakingChain } from 'consts/util'
 import { hexToU8a } from 'dedot/utils'
 import { useApi } from 'hooks/useApi'
 import { useNetwork } from 'hooks/useNetwork'
-import type { SignerPromptProps } from 'library/SubmitTx/types'
 import { useState } from 'react'
 import { useTranslation } from 'react-i18next'
+import type { SignerPromptProps } from 'types'
 import { ButtonPrimary, ButtonSecondary } from 'ui-buttons'
 import { usePrompt } from 'ui-overlay'
 import { Wrapper } from './Wrapper'
