--- packages/app-staking/src/library/SubmitTx/types.ts
@@ -1,46 +1,0 @@
-// Copyright 2026 @polkadot-cloud/polkadot-cloud-apps authors & contributors
-// SPDX-License-Identifier: GPL-3.0-only
-
-import type { VaultSignatureResult } from '@polkadot-cloud/connect-vault'
-import type {
-	ActiveAccount,
-	ActiveProxy,
-	DisplayFor,
-	MaybeAddress,
-} from 'types'
-
-export interface ProxySwitcherProps {
-	onPreviousSigner?: () => void
-	onNextSigner?: () => void
-	hasMultipleSigners?: boolean
-}
-
-export type SubmitTxProps = SubmitProps &
-	ProxySwitcherProps & {
-		requiresMigratedController?: boolean
-		proxySupported: boolean
-		noMargin?: boolean
-		onResize?: () => void
-		transparent?: boolean
-		txInitiated: boolean
-		proxyAccount: ActiveProxy | null
-		stacked?: boolean
-	}
-
-export interface SubmitProps {
-	uid: number
-	onSubmit: () => void
-	valid: boolean
-	submitText?: string
-	submitAccount: ActiveAccount
-	displayFor?: DisplayFor
-}
-
-export interface SignerPromptProps {
-	submitAddress: MaybeAddress
-	toSign: Uint8Array
-	onComplete: (
-		status: 'complete' | 'cancelled',
-		signature: VaultSignatureResult,
-	) => void
-}
